fix: naming and comments
